Change handler names to be consistent
